poc/image-clustering/pyproject.toml
[project]
name = "image-clustering-poc"
version = "0.1.0"
description = "POC for clustering visually similar images using CLIP embeddings"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "imgbeddings>=0.1.0",
    "hdbscan>=0.8.33",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "ty>=0.0.1a10",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.environment]
python-version = "3.11"
